ENH: Typescript 적용
