+++ package.json
@@ -15,15 +15,18 @@
     "lint:fix": "eslint . --fix"
   },
   "dependencies": {
+    "@types/styled-components": "^4.4.2",
     "@zeit/next-css": "^1.0.1",
     "next": "^9.1.5",
     "next-compose-plugins": "^2.2.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "semantic-ui-react": "^0.88.2",
-    "styled-components": "^5.0.1"
+    "styled-components": "^5.0.1",
+    "typescript": "^3.7.5"
   },
   "devDependencies": {
+    "@types/react": "^16.9.19",
     "babel-eslint": "^10.0.3",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
